Rename package
src -> vitae.
I think this looks better when calling from Python: py -m vitae.
