+++ vitae/features/ingestion/usecase.py
@@ -5,14 +5,14 @@ from __future__ import annotations
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
 
-from src.features.ingestion import scanners as strategy
-from src.features.ingestion.parsing import CurriculumParser
-from src.lib.panic import panic
+from vitae.features.ingestion import scanners as strategy
+from vitae.features.ingestion.parsing import CurriculumParser
+from vitae.lib.panic import panic
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from src.features.ingestion.repository import Researchers
+    from vitae.features.ingestion.repository import Researchers
 
 
 @dataclass(kw_only=True)
