+++ vitae/lib/functions.py
