+++ vitae/lib/panic.py
